# openapi.yaml
openapi: 3.0.1
info:
  title: Places Review Plugin
  description: A plugin that allows the user to retrieve personalized places based on a given address, radius, and type of place using ChatGPT.  It will return at most the top 10 recommendations sorted by closest distance. It will use user preferences to look through reviews and output most relevant places results. Ask at least 1 and at most 3 clarifying questions as needed to understand their preferences. Use deafult address of 1625 leavenworth st Apt 303 San Francisco, CA 94109 if not specified.
  version: 'v1'
servers:
  - url: http://localhost:3333
paths:
  /get_place_details/{address}/{radius}/{type_of}:
    get:
      operationId: getPlaceDetails
      summary: Retrieve reviews of places based on address, radius, and type of place
      parameters:
        - in: path
          name: address
          schema:
            type: string
          required: true
          description: The exact address to search for places around
        - in: path
          name: radius
          schema:
            type: integer
          required: true
          description: The radius (in meters) around the address to search for places
        - in: path
          name: type_of
          schema:
            type: string
          required: true
          description: The type of place to search for (e.g., restaurant, cafe, etc.)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: The name of the place
                    reviews:
                      type: array
                      items:
                        type: object
                        description: The reviews for the place
                    address:
                      type: string
                      description: The address of the place
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message


# openapi: 3.0.1
# info:
#   title: Place Reviews Plugin
#   description: A plugin that allows the user to retrieve reviews of places based on location, radius, and type using ChatGPT.
#   version: 'v1'
# servers:
#   - url: http://localhost:3333
# paths:
#   /places:
#     get:
#       operationId: getPlaceReviews
#       summary: Get reviews of places based on location, radius, and type
#       parameters:
#         - name: location
#           in: query
#           description: The latitude and longitude of the location around which to retrieve place information (e.g., "48.8588443,2.2943506").
#           required: true
#           schema:
#             type: string
#         - name: radius
#           in: query
#           description: The distance (in meters) within which to return place results.
#           required: true
#           schema:
#             type: integer
#         - name: type
#           in: query
#           description: The type of place to search for (e.g., "restaurant").
#           required: true
#           schema:
#             type: string
#       responses:
#         "200":
#           description: OK
#           content:
#             application/json:
#               schema:
#                 $ref: '#/components/schemas/getPlaceReviewsResponse'
#         "400":
#           description: Bad Request
#           content:
#             application/json:
#               schema:
#                 $ref: '#/components/schemas/errorResponse'
# components:
#   schemas:
#     getPlaceReviewsResponse:
#       type: object
#       properties:
#         reviews:
#           type: array
#           items:
#             type: object
#           description: The list of reviews for each place.
#     errorResponse:
#       type: object
#       properties:
#         error:
#           type: string
#           description: Error message.